moved getShortChannelId and getShortClaimId to class methods
