+++ models/claim.js
@@ -1,3 +1,29 @@
+const logger = require('winston');
+
+function sortResult (result, longId) {
+  let claimIndex;
+  let shortId = longId.substring(0, 1); // default sort id is the first letter
+  let shortIdLength = 0;
+  // find the index of this certificate
+  claimIndex = result.findIndex(element => {
+    return element.claimId === longId;
+  });
+  if (claimIndex < 0) { throw new Error('claimid not found in possible sorted list') }
+  // get an array of all certificates with lower height
+  let possibleMatches = result.slice(0, claimIndex);
+  // remove certificates with the same prefixes until none are left.
+  while (possibleMatches.length > 0) {
+    shortIdLength += 1;
+    shortId = longId.substring(0, shortIdLength);
+    possibleMatches = possibleMatches.filter(element => {
+      return (element.claimId.substring(0, shortIdLength) === shortId);
+    });
+  }
+  // return the short Id
+  logger.debug('short claim id ===', shortId);
+  return shortId;
+}
+
 module.exports = (sequelize, { STRING, BOOLEAN, INTEGER, TEXT, ARRAY, DECIMAL, DOUBLE }) => {
   const Claim = sequelize.define(
     'Claim',
@@ -151,6 +177,28 @@ module.exports = (sequelize, { STRING, BOOLEAN, INTEGER, TEXT, ARRAY, DECIMAL, D
       foreignKey: {
         allowNull: true,
       },
+    });
+  };
+
+  Claim.getShortClaimIdFromLongClaimId = function (claimId, claimName) {
+    return new Promise((resolve, reject) => {
+      logger.debug(`finding short claim id for ${claimId}#${claimId}`);
+      this
+        .findAll({
+          where: {name: claimName},
+          order: [['height', 'ASC']],
+        })
+        .then(result => {
+          switch (result.length) {
+            case 0:
+              throw new Error('That is an invalid claim name');
+            default:
+              resolve(sortResult(result, claimId));
+          }
+        })
+        .catch(error => {
+          reject(error);
+        });
     });
   };
