+++ passport/local-login.js
@@ -40,7 +40,7 @@ module.exports = new PassportLocalStrategy(
                 .then(channel => {
                   userInfo['channelName'] = channel.channelName;
                   userInfo['channelClaimId'] = channel.channelClaimId;
-                  return db.getShortChannelIdFromLongChannelId(channel.channelClaimId, channel.channelName);
+                  return db.Certificate.getShortChannelIdFromLongChannelId(channel.channelClaimId, channel.channelName);
                 })
                 .then(shortChannelId => {
                   userInfo['shortChannelId'] = shortChannelId;
